[IMPROVEMENT] Updating Koin to new DSL, compiler plugin, DSL

+++ app/src/main/java/com/brunodegan/ifood_challenge/data/datasources/local/entities/TopRatedMoviesEntity.kt
@@ -20,7 +20,7 @@ data class TopRatedMoviesEntity(
     @ColumnInfo(name = "popularity")
     val popularity: Double,
     @ColumnInfo(name = "vote_average")
-    val voteAverage: Double,
+    val voteAverage: Int,
     @ColumnInfo(name = "release_date")
     val releaseDate: String,
     @ColumnInfo(name = "is_favorite")
